Move aggregates: entities->system
