+++ tests/protocols/gmpv214/system/test_aggregates.py
@@ -17,7 +17,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from ...gmpv208 import Gmpv208TestCase
-from ...gmpv208.entities.aggregates import (
+from ...gmpv208.system.aggregates import (
     GmpGetAggregatesTestMixin,
 )
